Move internal modules into new _internal subpackage.
