+++ signac/_internal/dict_manager.py
