+++ signac/_internal/utility.py
