+++ signac/filterparse.py
@@ -6,7 +6,7 @@
 import json
 from collections.abc import Mapping
 
-from .utility import _print_err
+from ._internal.utility import _print_err
 
 
 def _is_json_like(q):
