+++ signac/h5store.py
@@ -10,9 +10,9 @@ import warnings
 from collections.abc import Mapping, MutableMapping
 from threading import RLock
 
-from .dict_manager import _DictManager
+from ._internal.dict_manager import _DictManager
+from ._internal.utility import _safe_relpath
 from .errors import H5StoreAlreadyOpenError, H5StoreClosedError, InvalidKeyError
-from .utility import _safe_relpath
 
 __all__ = [
     "H5Store",
