+++ signac/import_export.py
@@ -17,10 +17,10 @@ from string import Formatter
 from tempfile import TemporaryDirectory
 from zipfile import ZIP_DEFLATED, ZipFile
 
-from ._searchindexer import _SearchIndexer
+from ._internal.search_indexer import _SearchIndexer
+from ._internal.utility import _dotted_dict_to_nested_dicts, _mkdir_p
 from .errors import DestinationExistsError, StatepointParsingError
 from .job import Job
-from .utility import _dotted_dict_to_nested_dicts, _mkdir_p
 
 logger = logging.getLogger(__name__)
